+++ database/migrations/2022_06_24_073625_create_activity_outcome_histories_table.php
@@ -18,7 +18,7 @@ class CreateActivityOutcomeHistoriesTable extends Migration
             $table->date('date');
             $table->foreignId('outcome_id')->references('id')->on('activity_outcomes')->cascadeOnDelete();
             $table->foreignId('activity_id')->references('id')->on('activities')->cascadeOnDelete();
-            $table->integer('achievment');
+            $table->double('achievment');
             $table->string('file')->nullable();
             $table->foreignId('user_id')->nullable();
             $table->foreign('user_id')->references('id')->on('users')->nullOnDelete();
