+++ database/migrations/2022_06_24_073643_create_sub_activity_output_histories_table.php
@@ -18,7 +18,7 @@ class CreateSubActivityOutputHistoriesTable extends Migration
             $table->date('date');
             $table->foreignId('sub_activity_id')->references('id')->on('sub_activities');
             $table->foreignId('outcome_id')->references('id')->on('sub_activity_outputs')->cascadeOnDelete();
-            $table->integer('achievment');
+            $table->double('achievment');
             $table->string('file')->nullable();
             $table->foreignId('user_id')->nullable();
             $table->foreign('user_id')->references('id')->on('users')->nullOnDelete();
